Moved enum to seperate dir
