+++ src/enums/index.ts
@@ -1,0 +1,7 @@
+export enum UserStoreMethods {
+  getEmail = 'getEmail',
+  getToken = 'getToken',
+  setEmail = 'setEmail',
+  setToken = 'setToken',
+  isLoggedIn = 'isLoggedIn'
+}
