+++ src/store/modules/user.ts
@@ -1,16 +1,9 @@
 import Vue from 'vue'
 import Vuex, { StoreOptions } from 'vuex'
 import { UserState } from '@/models/user'
+import { UserStoreMethods } from '@/enums'
 
 Vue.use(Vuex)
-
-export enum UserStoreMethods {
-  getEmail = 'getEmail',
-  getToken = 'getToken',
-  setEmail = 'setEmail',
-  setToken = 'setToken',
-  isLoggedIn = 'isLoggedIn'
-}
 
 const TOKEN = 'token'
 const EMAIL = 'email'
