+++ src/views/login/content/LoginCard.vue
@@ -22,7 +22,7 @@ import { UsernameInput, PasswordInput } from '@/components/forms'
 import { username, password } from '@/components/validations/validationRules.ts'
 import { baseValidationRule } from '@/components/validations/baseValidationRule'
 import axios from 'axios'
-import { UserStoreMethods } from '@/store/modules/user'
+import { UserStoreMethods } from '@/enums'
 import { Action, Getter } from 'vuex-class'
 
 @Component({
